Revise csv write handler enable approach
  The previous approach appends csv write handler to the original
funcgraph handler. However, since funcgraph ent handler will
sneakingly consumes the ret record (if it is a leaf), causing the
ret record not handled by the funcgraph ret handler, so is my write
handler, the info of leaf will not be printed.

  Therefore, I decide to not use such "append" approach, but
replace the original handler with the csv_write one if csv_en.